Sletter bruker fra plattform når brukeren sletter seg selv

+++ public/js/settings.js
@@ -68,6 +68,69 @@ bekreftSlett.onclick = function () {
             }).then(() => {
                 //Sletter resterende bruker info -> Realtime database
                 firebase.database().ref("Bruker/" + whiz.Uid).remove().then(() => {
+                    
+                    firebase.database().ref("Steam gruppe/Medlemmer/"+whiz.Uid).remove().catch((error) => {
+                        console.log("Bruker er ikke steam medlem: " + error.message);
+                    }).then(() => {
+                        firebase.database().ref("Xbox gruppe/Medlemmer/"+whiz.Uid).remove().catch((error) => {
+                            console.log("Bruker er ikke xbox medlem: " + error.message);
+                        }).then(() => {
+                            firebase.database().ref("Playstation gruppe/Medlemmer/"+whiz.Uid).remove().catch((error) => {
+                                console.log("Bruker er ikke ps medlem: " + error.message);
+                            }).then(() => {
+                                firebase.database().ref("Switch gruppe/Medlemmer/"+whiz.Uid).remove().catch((error) => {
+                                    console.log("Bruker er ikke switch medlem: " + error.message);
+                                }).then(() => {
+                                    //Sletter eventuelt profilbilde brukeren har -> Storage 
+                                    firebase.storage().ref("user/" + whiz.Uid + "/profile.jpg").delete().catch((error) => {
+                                        //Error -> Brukeren har ikke profilbilde
+                                        console.log(error.message);
+                                    }).then(() => {
+                                        //Sletter auth-info til slutt -> Authentication
+                                        user.delete().then(() => {
+                                            //Bruker slettet
+                                            window.location = "/";
+                                        }).catch((error) => {
+                                            console.log(error.message);
+                                        });
+                                    });
+                                })
+                            }) 
+                        })
+                    })
+                });
+
+            });
+        }).catch((error) => {
+            window.alert(error.message);
+        })
+    } else {
+        window.alert("Bruker kansellerte");
+    }
+}
+
+/* Gammel slettBruker-funksjon. Ser finere ut, men sletter ikke brukeren fra plattformene
+bekreftSlett.onclick = function () {
+    var user = firebase.auth().currentUser;
+    const whiz = JSON.parse(sessionStorage.getItem("bruker"));
+    //var deletePrompt = prompt("Skriv inn passord");
+    var confirmPassword = passordBekreftSlettInput.value;
+
+    if (confirmPassword != null || confirmPassword != "") {
+        var credential = reautentiser(confirmPassword);
+        user.reauthenticateWithCredential(credential).then(() => {
+            firebase.database().ref("Bruker/" + whiz.Uid + "/Grupper").once('value', (snapshot) => {
+                if (snapshot.exists()) {
+                    snapshot.forEach(function (childSnap) {
+                        firebase.database().ref("Grupper/" + childSnap.key+"/Medlemmer/"+whiz.Uid).remove().catch((error) => {
+                            console.log(error.message);
+                            //!!!!!!!!!HER SKAL DET OGSÅ SLETTES EVT BILDE TIL GRUPPEN!!!!!!!!!
+                        });
+                    });
+                }
+            }).then(() => {
+                //Sletter resterende bruker info -> Realtime database
+                firebase.database().ref("Bruker/" + whiz.Uid).remove().then(() => {
                     //Sletter eventuelt profilbilde brukeren har -> Storage 
                     firebase.storage().ref("user/" + whiz.Uid + "/profile.jpg").delete().catch((error) => {
                         //Error -> Brukeren har ikke profilbilde
@@ -90,24 +153,4 @@ bekreftSlett.onclick = function () {
     } else {
         window.alert("Bruker kansellerte");
     }
-}
-
-
-/*
-//Sletter resterende bruker info -> Realtime database
-firebase.database().ref("Bruker/"+whiz.Uid).remove().then(() => {
-    //Sletter eventuelt profilbilde brukeren har -> Storage 
-    firebase.storage().ref("user/"+whiz.Uid+"/profile.jpg").delete().catch((error) => {
-        //Error -> Brukeren har ikke profilbilde
-        console.log(error.message);
-    }).then(() => {
-        //Sletter auth-info til slutt -> Authentication
-        user.delete().then(() => {
-            //Bruker slettet
-            window.location = "/";
-        }).catch((error) => {
-            console.log(error.message);
-        });
-    });
-});
-*/
+}*/
